CloseableAttachmentResult, GetAttachmentOperation

+++ attachment_name.go
@@ -4,7 +4,7 @@ type AttachmentName struct {
 	name        string `json:"Name"`
 	hash        string `json:"Hash"`
 	contentType string `json:"ContentType"`
-	size        int    `json:"Size"`
+	size        int64  `json:"Size"`
 }
 
 func (n *AttachmentName) getName() string {
@@ -30,10 +30,10 @@ func (n *AttachmentName) setContentType(contentType string) {
 	n.contentType = contentType
 }
 
-func (n *AttachmentName) getSize() int {
+func (n *AttachmentName) getSize() int64 {
 	return n.size
 }
 
-func (n *AttachmentName) setSize(size int) {
+func (n *AttachmentName) setSize(size int64) {
 	n.size = size
 }
